+++ src/main/java/net/thep2wking/oedldoedlresources/content/block/ingot/BlockDrunkium.java
@@ -31,7 +31,7 @@ public class BlockDrunkium extends ModBlockEffectBase {
 
 	public BlockDrunkium(String modid, String name, CreativeTabs tab, Material material, SoundType sound,
 			MapColor mapColor, int harvestLevel, ModToolTypes toolType, float hardness, float resistance,
-			float lightLevel, Potion effect, Potion effect2, Potion effect3, Potion effect4, boolean isDebuff) {
+			int lightLevel, Potion effect, Potion effect2, Potion effect3, Potion effect4, boolean isDebuff) {
 		super(modid, name, tab, material, sound, mapColor, harvestLevel, toolType, hardness, resistance, lightLevel,
 				effect, isDebuff);
 		this.effect2 = effect2;
