+++ src/main/java/net/thep2wking/oedldoedlresources/content/block/ingot/BlockSadistium.java
@@ -27,7 +27,7 @@ public class BlockSadistium extends ModBlockEffectBase {
 
 	public BlockSadistium(String modid, String name, CreativeTabs tab, Material material, SoundType sound,
 			MapColor mapColor, int harvestLevel, ModToolTypes toolType, float hardness, float resistance,
-			float lightLevel, Potion effect, Potion effect2, Potion effect3, boolean isDebuff) {
+			int lightLevel, Potion effect, Potion effect2, Potion effect3, boolean isDebuff) {
 		super(modid, name, tab, material, sound, mapColor, harvestLevel, toolType, hardness, resistance, lightLevel,
 				effect, isDebuff);
 		this.effect2 = effect2;
