+++ src/main/java/net/thep2wking/oedldoedlresources/content/block/ore/BlockEnderOre.java
@@ -20,7 +20,7 @@ import net.thep2wking.oedldoedlresources.init.ModItems;
 public class BlockEnderOre extends ModBlockOreBase {
 	public BlockEnderOre(String modid, String name, CreativeTabs tab, int minXp, int maxXp, Material material,
 			SoundType sound, MapColor mapColor, int harvestLevel, ModToolTypes toolType, float hardness,
-			float resistance, float lightLevel) {
+			float resistance, int lightLevel) {
 		super(modid, name, tab, minXp, maxXp, material, sound, mapColor, harvestLevel, toolType, hardness, resistance,
 				lightLevel);
 	}
